Renamed classes so they have "IRC" in front so they are not confused with Telegram.
