+++ lib/IrcHandlers/IrcActionHandler.js
