+++ lib/IrcHandlers/IrcErrorHandler.js
