+++ lib/IrcHandlers/IrcJoinHandler.js
