+++ lib/IrcHandlers/IrcKickHandler.js
