+++ lib/IrcHandlers/IrcMessageHandler.js
