+++ lib/IrcHandlers/IrcPartHandler.js
