+++ lib/TeleIrc.js
@@ -1,9 +1,9 @@
-const IrcActionHandler = require('./Irchandlers/ActionHandler');
-const IrcErrorHandler = require('./Irchandlers/ErrorHandler');
-const IrcJoinHandler = require('./IrcHandlers/JoinHandler');
-const IrcKickHandler = require('./Irchandlers/KickHandler');
-const IrcMessageHandler = require('./IrcHandlers/MessageHandler');
-const IrcPartHandler = require('./IrcHandlers/PartHandler');
+const IrcActionHandler = require('./Irchandlers/IrcActionHandler');
+const IrcErrorHandler = require('./Irchandlers/IrcErrorHandler');
+const IrcJoinHandler = require('./IrcHandlers/IrcJoinHandler');
+const IrcKickHandler = require('./Irchandlers/IrcKickHandler');
+const IrcMessageHandler = require('./IrcHandlers/IrcMessageHandler');
+const IrcPartHandler = require('./IrcHandlers/IrcPartHandler');
 
 const MessageRateLimiter = require("./MessageRateLimiter");
 const TeleIrcException = require('./TeleIrcException');
