+++ tests/IrcActionHandlerTests.js
@@ -1,17 +1,17 @@
 'use strict';
 
-const IrcMessageHandler = require("../lib/IrcHandlers/MessageHandler");
+const IrcActionHandler = require("../lib/IrcHandlers/IrcActionHandler");
 
 /**
  * Ensures that if the handler is disabled,
  * nothing happens.
  */
-exports.IrcMessageHandler_DisabledTest = function(assert) {
+exports.IrcActionHandler_DisabledTest = function(assert) {
     var message = undefined;
 
-    let uut = new IrcMessageHandler(undefined, false, (msg) => {message = msg;});
+    let uut = new IrcActionHandler(undefined, false, (msg) => {message = msg;});
 
-    uut.ReportMessage("User", "#channel", "Hello, World!");
+    uut.ReportAction("User", "#channel", "Did a thing!");
 
     // Disabled, message should remain undefined.
     assert.strictEqual(message, undefined);
@@ -23,7 +23,7 @@ exports.IrcMessageHandler_DisabledTest = function(assert) {
  * Ensures that if the handler is enabled,
  * its callback is activated, even if the black-list is undefined.
  */
-exports.IrcMessageHandler_UndefinedBlackListEnabledTest = function(assert) {
+exports.IrcActionHandler_UndefinedBlackListEnabledTest = function(assert) {
     DoSuccessTest(assert, undefined);
 
     assert.done();
@@ -33,7 +33,7 @@ exports.IrcMessageHandler_UndefinedBlackListEnabledTest = function(assert) {
  * Ensures that if the handler is enabled,
  * its callback is activated, even if the black-list is null.
  */
-exports.IrcMessageHandler_NullBlackListEnabledTest = function(assert) {
+exports.IrcActionHandler_NullBlackListEnabledTest = function(assert) {
     DoSuccessTest(assert, null);
 
     assert.done();
@@ -43,18 +43,18 @@ exports.IrcMessageHandler_NullBlackListEnabledTest = function(assert) {
  * Ensures that if the handler is enabled,
  * its callback is activated, if the user is not in the black-list.
  */
-exports.IrcMessageHandler_NamesNotMatchBlackListEnabledTest = function(assert) {
+exports.IrcActionHandler_NamesNotMatchBlackListEnabledTest = function(assert) {
     DoSuccessTest(assert, ["Someone"]);
 
     assert.done();
 };
 
-exports.IrcMessageHandler_BlackListNamesMatchTest = function(assert) {
+exports.IrcActionHandler_BlackListNamesMatchTest = function(assert) {
     var message = undefined;
 
-    let uut = new IrcMessageHandler(["user"], true, (msg) => {message = msg;});
+    let uut = new IrcActionHandler(["user"], true, (msg) => {message = msg;});
 
-    uut.ReportMessage("User", "#channel", "Hello, World!");
+    uut.ReportAction("User", "#channel", "Did a thing!");
 
     // Should be disabled, as a black-list name appeared.
     assert.strictEqual(message, undefined);
@@ -65,10 +65,10 @@ exports.IrcMessageHandler_BlackListNamesMatchTest = function(assert) {
 function DoSuccessTest(assert, blackList) {
     var message = undefined;
 
-    let uut = new IrcMessageHandler(blackList, true, (msg) => {message = msg;});
+    let uut = new IrcActionHandler(blackList, true, (msg) => {message = msg;});
 
-    uut.ReportMessage("User", "#channel", "Hello, World!");
+    uut.ReportAction("User", "#channel", "Did a thing!");
 
-    let expectedMessage = "<User> Hello, World!";
+    let expectedMessage = "User Did a thing!";
     assert.strictEqual(message, expectedMessage);
 };
